requirements.txt
dlib==19.24.4
h5py==3.8.0
matplotlib==3.5.3
numpy==1.21.5
opencv_contrib_python==3.4.17.61
opencv_python==4.7.0.72
pandas==1.3.5
Pillow==9.4.0
Pillow==10.4.0
pyevtk==1.6.0
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
Shapely==2.0.5
tabulate==0.9.0
tensorflow==2.10.0
yacs==0.1.8
